+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 3.0.0
 
 * Breaking change: Methods now return null on simple success (before they returned "OK").
+* Breaking change: Calling audioVideoStop() does not clear the view IDs anymore. Instead use clearViewIds().
 
 ## 2.0.3